Renaming src/common/types.ts
