+++ src/components/icons/product-icon/index.tsx
@@ -8,7 +8,7 @@ import { IconTerraformColor16 } from '@hashicorp/flight-icons/svg-react/terrafor
 import { IconVagrantColor16 } from '@hashicorp/flight-icons/svg-react/vagrant-color-16'
 import { IconVaultColor16 } from '@hashicorp/flight-icons/svg-react/vault-color-16'
 import { IconWaypointColor16 } from '@hashicorp/flight-icons/svg-react/waypoint-color-16'
-import { ProductSlug } from 'common/types'
+import { ProductSlug } from 'types/products'
 
 // TODO: is there a programatic way to build this from productNamesToIcons?
 interface ProductIconProps {
